public holiday manyLineParams

+++ src/database/migrations/2023_01_06_062122_create_public_holidays_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->integer('year');
             $table->unsignedBigInteger('workplace_id');
             $table->date('ph_date');
-            $table->integer('ph_hours');
+            $table->float('ph_hours');
 
             $table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
             $table->timestamp('updated_at')->default(DB::raw('CURRENT_TIMESTAMP on update CURRENT_TIMESTAMP'));;
